[Unstable] Switched to 64-bit color identifiers

+++ src/index/idQueue.cpp
@@ -1,7 +1,7 @@
 #include "idQueue.h"
 
 
-uint32_t IDQueue::pop()
+uint64_t IDQueue::pop()
 {
     if (queue.size()==0)
     {
@@ -9,12 +9,12 @@ uint32_t IDQueue::pop()
         queue.push_back(max_multiplier * bins + offset);
     }
 
-    uint32_t id = *queue.begin();
+    uint64_t id = *queue.begin();
     queue.erase(queue.begin());
     return id;
 }
 
-void IDQueue::push(uint32_t id)
+void IDQueue::push(uint64_t id)
 {
     queue.push_back(id);
 }
