+++ src/index/idQueue.h
@@ -8,12 +8,12 @@ class IDQueue
 {
 
     protected:
-        uint32_t max_multiplier = 0b0u;
-        vector<uint32_t> queue;
+        uint64_t max_multiplier = 0b0u;
+        vector<uint64_t> queue;
 
     public:
-        uint32_t bins;
-        uint32_t offset;
-        uint32_t pop();
-        void push(uint32_t id);
+        uint64_t bins;
+        uint64_t offset;
+        uint64_t pop();
+        void push(uint64_t id);
 };
